Changes to compile with MSVC. Not that printer and printers classes do not retrieve printer info from Windows. Also, no tests are compiled for MSVC.

+++ include/pageorientation.h
@@ -20,7 +20,6 @@
 #include <iostream>
 #include <vector>
 #include <optional>
-#include <cups/cups.h>
 #include "values.h"
 
 namespace nanaprint
@@ -34,13 +33,13 @@ namespace nanaprint
     {
         public:
             page_orientation();
-            page_orientation(int orientation);
+            page_orientation(size_t orientation);
             virtual ~page_orientation() {}
             const std::optional<std::string>& get_orientation() const;
-            int get_orientation_number() const;
+            size_t get_orientation_number() const;
 
         private:
-            int m_orientationNum;
+            size_t m_orientationNum;
             std::optional<std::string> m_orientation;
     };
