+++ include/printsettings.h
@@ -31,7 +31,7 @@ namespace nanaprint
             virtual ~print_settings() {}
             const printers& get_printers() const
                 { return m_printers; }
-            void set_printer(int printer);
+            void set_printer(size_t printer);
             size_t get_printer() const { return m_printer; }
             void set_borderless(bool borderless);
             bool get_borderless() const { return m_borderless; }
@@ -54,11 +54,11 @@ namespace nanaprint
             const std::optional<side>& get_side() const { return m_side; }
             bool can_print_multiple_copies() const { return m_canPrintMultipleCopies; }
         protected:
-            int get_default_printer_number() const;
-            void set_default_settings(int printerNum);
+            size_t get_default_printer_number() const;
+            void set_default_settings(size_t printerNum);
 
             printers m_printers;
-            int m_printer;
+            size_t m_printer;
             bool m_borderless;
             std::optional<media_size> m_mediaSize;
             finishings m_finishings;
